[parser] Add parsing of assignment expressions
Note that `Expr::clone` was introduced in this commit as a
way to copy ASTs.  This was needed for cases where one would
normally share pointers, but since this AST implementation
uses `std::unique_ptr` pointers can't be shared.  Instead
of moving everything to a `std::shared_ptr` for the minority
case where nodes are shared, a copy via `clone` is done instead.

Interpretation of assignment will be added in a later commit.